+++ lang/fr/about.php
@@ -135,6 +135,10 @@ return [
                 'box' => 'BP 12525',
                 'phone' => '+237 653 283 107',
                 'email' => 'gsautosbilan@gmail.com',
+                'emails' => [
+                    'gsautosbilan@gmail.com',
+                    'admin@gsautobilan.com',
+                ],
             ],
         ],
     ],